third_party/blink/web_tests/platform/generic/fast/multicol/caret-range-anonymous-block-expected.txt

Test that hitting an anonymous block child of a multicol container takes us to the right text content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

First column:
PASS document.caretRangeFromPoint(170,39).startOffset is 3
PASS document.caretRangeFromPoint(170,40).startOffset is 7
PASS document.caretRangeFromPoint(170,79).startOffset is 11
Second column:
PASS document.caretRangeFromPoint(270,39).startOffset is 15
PASS document.caretRangeFromPoint(270,40).startOffset is 19
PASS document.caretRangeFromPoint(270,79).startOffset is 23
Third column:
PASS document.caretRangeFromPoint(370,39).startOffset is 27
PASS document.caretRangeFromPoint(370,40).startOffset is 31
PASS document.caretRangeFromPoint(370,79).startOffset is 35
PASS successfullyParsed is true

TEST COMPLETE

xxx xxx xxx xxx xxx xxx xxx xxx xxx
